fixed: use opengl stipple patterns to render interlaced content weaved
Currently the weave pattern is not timingwise aligned to anything, thus if you use an interlaced output there is no guarantee that the correct field in time (of the time and space continum), will be displayed first on your interlaced display. It could just aswell happen that the fields start displaying in opposite order which will create a very uggly experience. It can temporarily be solved by switching to the inverted mode, but that doesn't neccessarily last the whole movie. We'd need to sync the start of rendering to when the display is rendering it's correct field.