added: libdvdnav patch from upstream mplayer rev1239
Always attempt to move forward on a seek

From John Stebbins original post (with pathc, Thanks!):
libdvdnav has a problem in that it is difficult to gracefully recover
from a read error and continue on to subsequent blocks on the disc.  An
application would like to seek forward past the current block after
getting a read error in order to attempt to get past the bad block(s).
But dvdnav_sector_search() does not guarantee that a requested forward
seek will actually move the current position forward.  It truncates down
to the nearest VOBU which will almost always cause a forward seek
request by a single block (or a small number of blocks) to move the
current position backward.  This behaviour puts the application into a
loop of: read failure, attempted forward seek (which results in backward
seek), read failure ...