[b95b457] [fixed] Yet Another Race Condition Fix
This fixes the race condition between SendMessage and ProcessMessages/ProcessWindowsMessages.

To the best of my knowledge this is was also the issue with vrdfan's 32bit ripping problem. I managed to reproduce it on a single processor 32 bit linux box myself and this addresses the issue - though it's possible I didn't reproduce the same issue.

As far as I can tell this is the last known remaining issue caused by, or uncovered by, the threading refactor.